+++ query-planner-js/src/__tests__/buildPlan.test.ts
@@ -4052,6 +4052,88 @@ describe('Named fragments preservation', () => {
       }
     `);
   });
+
+  it('do not try to apply fragments that are not valid for the subgaph', () => {
+    // Slightly artificial example for simplicity, but this highlight the problem.
+    // In that example, the only queried subgraph is the first one (there is in fact
+    // no way to ever reach the 2nd one), so the plan should mostly simply forward
+    // the query to the 1st subgraph, but a subtlety is that the named fragment used
+    // in the query is *not* valid for Subgraph1, because it queries `b` on `I`, but
+    // there is no `I.b` in Subgraph1.
+    // So including the named fragment in the fetch would be erroneous: the subgraph
+    // server would reject it when validating the query, and we must make sure it
+    // is not reused.
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          i1: I
+          i2: I
+        }
+
+        interface I {
+          a: Int
+        }
+
+        type T implements I {
+          a: Int
+          b: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        interface I {
+          a: Int
+          b: Int
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      query {
+        i1 {
+          ... on T {
+            ...Frag
+          }
+        }
+        i2 {
+          ... on T {
+            ...Frag
+          }
+        }
+      }
+
+      fragment Frag on I {
+        b
+      }
+    `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            i1 {
+              __typename
+              ... on T {
+                b
+              }
+            }
+            i2 {
+              __typename
+              ... on T {
+                b
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
 });
 
 test('works with key chains', () => {
@@ -4584,5 +4666,595 @@ describe('interface type-explosion', () => {
       }
     `);
     expect(queryPlanner.lastGeneratedPlanStatistics()?.evaluatedPlanCount).toBe(1);
+  });
+});
+
+/*
+ * Those tests the cases where 2 abstract types (interface or union) interact (having some common runtime
+ * types intersection), but one of them include an runtime type that the other also include _in the supergraph_
+ * but *not* in one of the subgraph. The tl;dr is that in some of those interaction, we must force a type-explosion
+ * to handle it properly, but no in other interactions, and this ensures this is handled properly.
+ */
+describe('merged abstract types handling', () => {
+  test('union/interface interaction', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          u: U
+        }
+
+        union U = A | B | C
+
+        interface I {
+          v: Int
+        }
+
+        type A {
+          v: Int @shareable
+        }
+
+        type B implements I {
+          v: Int
+        }
+
+        type C implements I {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        interface I {
+          v: Int
+        }
+
+        type A implements I {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        u {
+          ... on I {
+            v
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // Type `A` can be returned by `u` and is a `I` *in the supergraph* but not in `Subgraph1`, so need to
+    // type-explode `I` in the query to `Subgraph1` so it doesn't exclude `A`.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            u {
+              __typename
+              ... on A {
+                v
+              }
+              ... on B {
+                v
+              }
+              ... on C {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('union/interface interaction, but no need to type-explode', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          u: U
+        }
+
+        union U = B | C
+
+        interface I {
+          v: Int
+        }
+
+        type A implements I {
+          v: Int @shareable
+        }
+
+        type B implements I {
+          v: Int
+        }
+
+        type C implements I {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        union U = A
+
+        type A {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        u {
+          ... on I {
+            v
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // While `A` is a `U` in the supergraph while not in `Subgraph1`, since the `u`
+    // operation is resolved by `Subgraph1`, it cannot ever return a A, and so
+    // there is need to type-explode `I` in this query.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            u {
+              __typename
+              ... on I {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('interface/union interaction', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          i: I
+        }
+
+        union U = B | C
+
+        interface I {
+          v: Int
+        }
+
+        type A implements I {
+          v: Int @shareable
+        }
+
+        type B implements I {
+          v: Int
+        }
+
+        type C implements I {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        union U = A
+
+        type A {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        i {
+          ... on U {
+            ... on A {
+              v
+            }
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // Type `A` can be returned by `i` and is a `U` *in the supergraph* but not in `Subgraph1`, so need to
+    // type-explode `U` in the query to `Subgraph1` so it doesn't exclude `A`.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            i {
+              __typename
+              ... on A {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('interface/union interaction, but no need to type-explode', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          i: I
+        }
+
+        union U = A | B | C
+
+        interface I {
+          v: Int
+        }
+
+        type A {
+          v: Int @shareable
+        }
+
+        type B implements I {
+          v: Int
+        }
+
+        type C implements I {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        interface I {
+          v: Int
+        }
+
+        type A implements I {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        i {
+          ... on U {
+            ... on A {
+              v
+            }
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // While `A` is a `I` in the supergraph while not in `Subgraph1`, since the `i` operation is resolved by
+    // `Subgraph1`, it cannot ever return a A, and so we should skip the whole `v` selection; or at the very
+    // least, we should not send a query with `... on U { ... on A { <stuff> }}` to `Subgraph1` since it
+    // would reject it as invalid.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            i {
+              __typename
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('interface/interface interaction', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          i1: I1
+        }
+
+        interface I1 {
+          v: Int
+        }
+
+        interface I2 {
+          v: Int
+        }
+
+        type A implements I1 {
+          v: Int @shareable
+        }
+
+        type B implements I1 & I2 {
+          v: Int
+        }
+
+        type C implements I1 & I2 {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        interface I2 {
+          v: Int
+        }
+
+        type A implements I2 {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        i1 {
+          ... on I2 {
+            v
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // Type `A` can be returned by `i1` and is a `I2` *in the supergraph* but not in `Subgraph1`, so need to
+    // type-explode `I2` in the query to `Subgraph1` so it doesn't exclude `A`.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            i1 {
+              __typename
+              ... on A {
+                v
+              }
+              ... on B {
+                v
+              }
+              ... on C {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('interface/interface interaction, but no need to type-explode', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          i1: I1
+        }
+
+        interface I1 {
+          v: Int
+        }
+
+        interface I2 {
+          v: Int
+        }
+
+        type A implements I2 {
+          v: Int @shareable
+        }
+
+        type B implements I1 & I2 {
+          v: Int
+        }
+
+        type C implements I1 & I2 {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        interface I1 {
+          v: Int
+        }
+
+        type A implements I1 {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        i1 {
+          ... on I2 {
+            v
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // While `A` is a `I1` in the supergraph while not in `Subgraph1`, since the `i1`
+    // operation is resolved by `Subgraph1`, it cannot ever return a A, and so
+    // there is need to type-explode `I2` in this query (even if `Subgraph1` would
+    // otherwise not include `A` from a `... on I2`).
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            i1 {
+              __typename
+              ... on I2 {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  test('union/union interaction', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          u1: U1
+        }
+
+        union U1 = A | B | C
+        union U2 = B | C
+
+        type A {
+          v: Int @shareable
+        }
+
+        type B {
+          v: Int
+        }
+
+        type C {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        union U2 = A
+
+        type A {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        u1 {
+          ... on U2 {
+            ... on A {
+              v
+            }
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // Type `A` can be returned by `u1` and is a `U2` *in the supergraph* but not in `Subgraph1`, so need to
+    // type-explode `U2` in the query to `Subgraph1` so it doesn't exclude `A`.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            u1 {
+              __typename
+              ... on A {
+                v
+              }
+            }
+          }
+        },
+      }
+    `);
+  });
+
+  // TODO: this test currently doesn't work due to https://github.com/apollographql/federation/issues/2256
+  // (it is not a direct test of that issue, but one of its consequence nonetheles). We should enable it
+  // with the fix of that issue.
+  test.skip('union/union interaction, but no need to type-explode', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        type Query {
+          u1: U1
+        }
+
+        union U1 = B | C
+        union U2 = A | B | C
+
+        type A {
+          v: Int @shareable
+        }
+
+        type B {
+          v: Int
+        }
+
+        type C {
+          v: Int
+        }
+      `
+    }
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        union U1 = A
+
+        type A {
+          v: Int @shareable
+        }
+      `
+    }
+
+    const [api, queryPlanner] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const operation = operationFromDocument(api, gql`
+      {
+        u1 {
+          ... on U2 {
+            ... on A {
+              v
+            }
+          }
+        }
+      }
+      `);
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    // While `A` is a `U1` in the supergraph while not in `Subgraph1`, since the `u1` operation is resolved by
+    // `Subgraph1`, it cannot ever return a A, and so we should skip the whole `v` selection; or at the very
+    // least, we should not send a query with `u1 { ... on A { <stuff> }}` to `Subgraph1` since it
+    // would reject it as invalid.
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            u1 {
+              __typename
+            }
+          }
+        },
+      }
+    `);
   });
 });